Ajout Rules & Warn

+++ major/major.py
@@ -119,7 +119,7 @@ class MajorAPI:
         flammes = len(data["DATA"]["flammes"])
         der_msg = data["DATA"]["flammes"][-1] if flammes else ajd
         xp = (int(fmsg_jours) + int((datetime.now() - user.joined_at).days / 5)) * data["DATA"]["msg_nb"]
-        xp = xp/(datetime.now() - user.server.created_at).days
+        xp = int(xp/(datetime.now() - user.server.created_at).days)
         logs = data["LOGS"][::-1]
         pseudos, surnoms = data["DATA"]["pseudos"][::-1], data["DATA"]["surnoms"][::-1]
         StatsData = namedtuple('StatsData', ['msg_nb', 'msg_suppr', 'emojis', 'join', 'quit', 'pseudos', 'surnoms',
@@ -141,7 +141,7 @@ class MajorAPI:
                     firstmsg = datetime.fromtimestamp(data["DATA"]["first_msg"])
                     fmsg_jours = (datetime.now() - firstmsg).days
                     xp = (int(fmsg_jours) + int((datetime.now() - user.joined_at).days / 5)) * data["DATA"]["msg_nb"]
-                    xp = xp/(datetime.now() - server.created_at).days
+                    xp = int(xp/(datetime.now() - server.created_at).days)
                     liste.append([xp, u])
                 except:
                     pass
